simplify deployment
